Update deploy scripts. Configs and mods working.

+++ deploy/deploy_configs.sh
@@ -8,10 +8,14 @@ SFTP_HOST="192.168.68.103"
 SFTP_PORT="2224"
 SFTP_USER="update"
 
-echo "Please enter password:"
-read SFTP_PASS
+source deploy.env
 
-echo "Deleting configs on the remote server"
+if [ -z ${SFTP_PASS+x} ]; then
+    echo "Could not find SFTP_PASS in deploy.env"
+    exit 1 
+fi
+
+echo "Deploying configs to server."
 
 sshpass -p "$SFTP_PASS" ssh -p $SFTP_PORT $SFTP_USER@$SFTP_HOST "rm -f config/*"
 
@@ -23,7 +27,7 @@ else
 fi
 
 echo "Uploading local configs to config folder"
-sshpass -p "$SFTP_PASS" scp -P $SFTP_PORT ../config/* $SFTP_USER@$SFTP_HOST:config/
+sshpass -p "$SFTP_PASS" scp -P $SFTP_PORT -r ../config/* $SFTP_USER@$SFTP_HOST:config/
 if [ $? -eq 0 ] || [ $? -eq 1 ]; then
     echo "Successfully uploaded configs."
 else
